pyproject.toml
[tool.poetry]
name = "nullshift"
version = "0.1.0"
description = "AI-powered patchflow agent that detects untested Python functions on pull requests and autonomously generates unit tests."
authors = ["Ravikiranreddybada"]
readme = "README.md"
packages = [
    { include = "patchwork", from = "." }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.9"
typing-extensions = "^4.11.0"
click = "~8.1.0"
openai = "^1.46.1"
tree-sitter-languages = "~1.10.2"
python-gitlab = "^4.13.0"
pygithub = "~2.1.1"
azure-devops = "~7.1.0b4"
gitpython = "~3.1.40"
pydantic = "~2.10.6"
pydantic-ai = "^0.0.37"
google-cloud-aiplatform = "^1.84.0"
google-genai = "^1.2.0"
libcst = "~1.1.0"
sqlalchemy = "~2.0.36"
patched-code2prompt = "~0.9.0.dev1"
pyyaml = "^6.0.1"
packageurl-python = "~0.15.0"
semver = "~3.0.2"
requests = "~2.32.3"
chardet = "~5.2.0"
attrs = "~23.2.0"
anthropic = "^0.49.0"
rich = "~13.7.1"
chevron = "~0.14.0"
giturlparse = "~0.12.0"
eml_parser = "^2.0.0"
csvkit = "^2.1.0"
python-magic = "^0.4.27"
scikit-learn = "^1.3.2"
json-repair = "~0.30.0"
dotenv = "^0.9.9"
# pinning transitive dependencies
tree-sitter = "~0.21.3"
numpy = "1.26.4"
pandas = "2.2.3"
scipy = "1.13.1"
boto3 = "~1.37.11"

# extras
# security
semgrep = { version = "1.52.0", optional = true }
owasp-depscan = { version = "5.2.15", optional = true }

# notifications
slack-sdk = { version = "~3.28.0", optional = true }

# browser-use
browser-use = [{ version = "~0.1.40", python = ">=3.11", optional = true }]
playwright = [{ version = "~1.50.0", optional = true }]
langchain-google-genai = [{ version = "^2.0.10", optional = true }]
langchain-openai = [{ version = "~0.3.1", optional = true }]
langchain-anthropic = [{ version = "~0.3.3", optional = true }]
pillow = [{ version = "~10.4.0", optional = true }]

[tool.poetry.extras]
security = ["semgrep", "owasp-depscan"]
notifications = ["slack-sdk"]
browser-use = ["browser-use", "playwright", "langchain-google-genai", "langchain-openai", "langchain-anthropic", "Pillow"]
all = [
    "semgrep", "owasp-depscan",
    "slack-sdk",
    "browser-use", "playwright", "langchain-google-genai", "langchain-openai", "langchain-anthropic", "Pillow"
]

[tool.poetry.group.dev.dependencies]
setuptools = "*"
poethepoet = {version = "^0.33.1", extras = ["poetry-plugin"]}
mypy = "^1.7.1"
types-requests = "~2.31.0"
black = "^23.12.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
pytest = "^8.1.1"
pytest-mock = "^3.8.0"
pytest-localserver = "^0.8.1"
datasets = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
nullshift = 'patchwork.app:cli'

[tool.poe.poetry_hooks]
pre_build = "mypy"
post_update = "lint"

[tool.poe.env]
PROJ_PATH.default = "patchwork"
TEST_PATH.default = "tests"

[tool.poe.tasks]
mypy = "mypy patchwork --ignore-missing-imports"
lint = [
    { cmd = "autoflake --recursive ${PROJ_PATH} ${TEST_PATH}" },
    { cmd = "isort ${PROJ_PATH} ${TEST_PATH}" },
    { cmd = "black ${PROJ_PATH} ${TEST_PATH}" }
]

[tool.black]
target-version = ['py310']
line-length = 120

[tool.isort]
profile = "black"

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
expand-star-imports = true
ignore-init-module-imports = true
